ST-1472 DSS submit backend
