+++ src/main/java/uk/gov/hmcts/sptribs/citizen/event/CicCreateCaseEvent.java
@@ -14,8 +14,8 @@ import uk.gov.hmcts.sptribs.common.ccd.CcdCaseType;
 import uk.gov.hmcts.sptribs.common.config.AppsConfig;
 import uk.gov.hmcts.sptribs.util.AppsUtil;
 
-import java.util.ArrayList;
-
+import static uk.gov.hmcts.sptribs.ciccase.model.UserRole.CITIZEN_CIC;
+import static uk.gov.hmcts.sptribs.ciccase.model.UserRole.CREATOR;
 import static uk.gov.hmcts.sptribs.ciccase.model.access.Permissions.CREATE_READ_UPDATE;
 
 @Component
@@ -27,18 +27,14 @@ public class CicCreateCaseEvent implements CCDConfig<CaseData, State, UserRole> 
 
     @Override
     public void configure(final ConfigBuilder<CaseData, State, UserRole> configBuilder) {
-        var defaultRoles = new ArrayList<UserRole>();
-        defaultRoles.add(UserRole.CITIZEN_CIC);
-
-        var updatedRoles = defaultRoles;
 
         configBuilder
             .event(AppsUtil.getExactAppsDetailsByCaseType(appsConfig, CcdCaseType.CIC.getCaseTypeName()).getEventIds()
                 .getCreateEvent())
             .initialState(State.Draft)
-            .name("Create draft case (cic)")
-            .description("Apply for edge case (cic)")
-            .grant(CREATE_READ_UPDATE, updatedRoles.toArray(UserRole[]::new))
+            .name("Create draft case (DSS)")
+            .description("Apply for edge case (DSS)")
+            .grant(CREATE_READ_UPDATE, CITIZEN_CIC, CREATOR)
             .aboutToSubmitCallback(this::aboutToSubmit)
             .retries(120, 120);
     }
@@ -46,9 +42,10 @@ public class CicCreateCaseEvent implements CCDConfig<CaseData, State, UserRole> 
     public AboutToStartOrSubmitResponse<CaseData, State> aboutToSubmit(CaseDetails<CaseData, State> details,
                                                                        CaseDetails<CaseData, State> beforeDetails) {
         var caseData = details.getData();
+        caseData.setHyphenatedCaseRef(details.getData().formatCaseRef(details.getId()));
         return AboutToStartOrSubmitResponse.<CaseData, State>builder()
             .data(caseData)
-            .state(State.DSS_Submitted)
+            .state(State.DSS_Draft)
             .build();
     }
