+++ src/test/java/uk/gov/hmcts/sptribs/citizen/event/CicCreateCaseEventTest.java
@@ -1,7 +1,5 @@
 package uk.gov.hmcts.sptribs.citizen.event;
 
-import com.google.common.collect.ImmutableSetMultimap;
-import com.google.common.collect.SetMultimap;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -12,33 +10,21 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.TestPropertySource;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
-import uk.gov.hmcts.ccd.sdk.ConfigBuilderImpl;
 import uk.gov.hmcts.ccd.sdk.api.CaseDetails;
-import uk.gov.hmcts.ccd.sdk.api.Event;
-import uk.gov.hmcts.ccd.sdk.api.Permission;
 import uk.gov.hmcts.ccd.sdk.api.callback.AboutToStartOrSubmitResponse;
 import uk.gov.hmcts.sptribs.ciccase.model.CaseData;
 import uk.gov.hmcts.sptribs.ciccase.model.State;
-import uk.gov.hmcts.sptribs.ciccase.model.UserRole;
 import uk.gov.hmcts.sptribs.common.AddSystemUpdateRole;
 import uk.gov.hmcts.sptribs.common.config.AppsConfig;
 import uk.gov.hmcts.sptribs.constants.CommonConstants;
-import uk.gov.hmcts.sptribs.util.AppsUtil;
 
-import java.util.Arrays;
 import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.mockito.ArgumentMatchers.anyList;
-import static org.mockito.Mockito.when;
-import static uk.gov.hmcts.ccd.sdk.api.Permission.C;
-import static uk.gov.hmcts.ccd.sdk.api.Permission.R;
-import static uk.gov.hmcts.ccd.sdk.api.Permission.U;
-import static uk.gov.hmcts.sptribs.ciccase.model.UserRole.CITIZEN_CIC;
 import static uk.gov.hmcts.sptribs.constants.CommonConstants.ST_CIC_CASE_TYPE;
-import static uk.gov.hmcts.sptribs.testutil.ConfigTestUtil.createCaseDataConfigBuilder;
-import static uk.gov.hmcts.sptribs.testutil.ConfigTestUtil.getEventsFrom;
 import static uk.gov.hmcts.sptribs.testutil.TestConstants.CASE_DATA_CIC_ID;
+import static uk.gov.hmcts.sptribs.testutil.TestConstants.TEST_CASE_ID;
+import static uk.gov.hmcts.sptribs.testutil.TestDataHelper.caseData;
 
 
 @ExtendWith(SpringExtension.class)
@@ -75,69 +61,13 @@ class CicCreateCaseEventTest {
     }
 
     @Test
-    void shouldAddConfigurationToConfigBuilderAndSetPermissionOnlyForCitizenRole() {
-        final ConfigBuilderImpl<CaseData, State, UserRole> configBuilder = createCaseDataConfigBuilder();
-
-        when(addSystemUpdateRole.addIfConfiguredForEnvironment(anyList()))
-            .thenReturn(List.of(CITIZEN_CIC));
-
-        when(appsConfig.getApps()).thenReturn(Arrays.asList(cicAppDetail));
-
-        cicCreateCaseEvent.configure(configBuilder);
-
-        assertThat(getEventsFrom(configBuilder).values())
-            .extracting(Event::getId)
-            .contains(AppsUtil.getExactAppsDetailsByCaseType(appsConfig, ST_CIC_CASE_TYPE).getEventIds()
-                          .getCreateEvent());
-
-        SetMultimap<UserRole, Permission> expectedRolesAndPermissions =
-            ImmutableSetMultimap.<UserRole, Permission>builder()
-                .put(CITIZEN_CIC, C)
-                .put(CITIZEN_CIC, R)
-                .put(CITIZEN_CIC, U)
-                .build();
-
-        assertThat(getEventsFrom(configBuilder).values())
-            .extracting(Event::getGrants)
-            .containsExactly(expectedRolesAndPermissions);
-    }
-
-    @Test
-    void shouldSetPermissionForCitizenAndCaseWorkerRoleWhenEnvironmentIsAat() {
-        final ConfigBuilderImpl<CaseData, State, UserRole> configBuilder = createCaseDataConfigBuilder();
-
-        when(addSystemUpdateRole.addIfConfiguredForEnvironment(anyList()))
-            .thenReturn(List.of(CITIZEN_CIC));
-
-        when(appsConfig.getApps()).thenReturn(Arrays.asList(cicAppDetail));
-
-        cicCreateCaseEvent.configure(configBuilder);
-
-        assertThat(getEventsFrom(configBuilder).values())
-            .extracting(Event::getId)
-            .contains("citizen-cic-create-dss-application");
-
-        assertThat(getEventsFrom(configBuilder).values())
-            .extracting(Event::getDescription)
-            .contains("Apply for edge case (cic)");
-
-        SetMultimap<UserRole, Permission> expectedRolesAndPermissions =
-            ImmutableSetMultimap.<UserRole, Permission>builder()
-                .put(CITIZEN_CIC, C)
-                .put(CITIZEN_CIC, R)
-                .put(CITIZEN_CIC, U)
-                .build();
-
-        assertThat(getEventsFrom(configBuilder).values())
-            .extracting(Event::getGrants)
-            .containsExactlyInAnyOrder(expectedRolesAndPermissions);
-    }
-
-    @Test
     void shouldChangeCaseStateWhenAboutToSubmit() {
         // Given
         CaseDetails<CaseData, State> details = new CaseDetails<>();
         CaseDetails<CaseData, State> beforeDetails = new CaseDetails<>();
+        details.setId(TEST_CASE_ID);
+        CaseData data = caseData();
+        details.setData(data);
 
         // When
         AboutToStartOrSubmitResponse<CaseData, State> response = cicCreateCaseEvent.aboutToSubmit(
@@ -146,6 +76,6 @@ class CicCreateCaseEventTest {
         );
 
         // Then
-        assertThat(response.getState()).isEqualTo(State.DSS_Submitted);
+        assertThat(response.getState()).isEqualTo(State.DSS_Draft);
     }
 }
